requirements.txt
telethon
python-binance
python-dotenv
requests
schedule
gspread
oauth2client
apscheduler
pytz
